Separate RightPane into own class
